requirements.txt
discord.py
discord-py-interactions
python-dotenv
arrow
